test: fix snpshot frequent timeout in CI
