+++ tests/snapshot/fail.test.ts
@@ -1,0 +1,26 @@
+import { dirname } from 'node:path';
+import { fileURLToPath } from 'node:url';
+import { expect, it } from '@rstest/core';
+import { runRstestCli } from '../scripts';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+it('should failed when snapshot unmatched', async () => {
+  const { cli } = await runRstestCli({
+    command: 'rstest',
+    args: ['run', 'fixtures/fail.test.ts'],
+    options: {
+      nodeOptions: {
+        cwd: __dirname,
+      },
+    },
+  });
+
+  await cli.exec;
+  expect(cli.exec.process?.exitCode).toBe(1);
+
+  const logs = cli.stdout.split('\n').filter(Boolean);
+
+  expect(logs.find((log) => log.includes('Snapshots 1 failed'))).toBeTruthy();
+});
